+++ Scripts/openapi_generate.sh
@@ -656,12 +656,9 @@ inject_v1_endpoint_paths() {
     case reminder(MessageId)
 
     case banMember
-    case flagUser(Bool)
-    case flagMessage(Bool)
+    case flagUser
+    case flagMessage
     case muteUser(Bool)
-
-    case callToken(String)
-    case createCall(String)
 
 EOF
 
@@ -709,11 +706,9 @@ EOF
         case let .reminder(messageId): return "messages/\(messageId)/reminders"
 
         case .banMember: return "moderation/ban"
-        case let .flagUser(flag): return "moderation/\(flag ? "flag" : "unflag")"
-        case let .flagMessage(flag): return "moderation/\(flag ? "flag" : "unflag")"
+        case .flagUser: return "moderation/flag"
+        case .flagMessage: return "moderation/flag"
         case let .muteUser(mute): return "moderation/\(mute ? "mute" : "unmute")"
-        case let .callToken(callId): return "calls/\(callId)"
-        case let .createCall(queryString): return "channels/\(queryString)/call"
 
 EOF